% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.xtdhcoint}
\alias{coef.xtdhcoint}
\title{Coef Method for xtdhcoint Objects}
\usage{
\method{coef}{xtdhcoint}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"xtdhcoint"}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Named numeric vector of z-statistics and p-values.
}
\description{
Extracts the test statistics from the Durbin-Hausman test.
}
