% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtdhcoint-package.R
\docType{package}
\name{xtdhcoint-package}
\alias{xtdhcoint-package}
\title{xtdhcoint: Durbin-Hausman Panel Cointegration Tests}
\description{
Implements the Durbin-Hausman panel cointegration tests of Westerlund (2008) \doi{10.1002/jae.967}. The tests are robust to cross-sectional dependence through common factor extraction using principal components. Provides both group-mean (DHg) and panel (DHp) test statistics with automatic factor number selection via information criteria.

Implements the Durbin-Hausman panel cointegration tests of
Westerlund (2008). The tests are robust to cross-sectional dependence
through common factor extraction using principal components.
}
\details{
The package provides the \code{\link{xtdhcoint}} function for performing
panel cointegration tests that account for cross-sectional dependence.

The main features include:
\itemize{
  \item DHg (group-mean) statistic for heterogeneous cointegration
  \item DHp (panel) statistic for homogeneous cointegration
  \item Automatic factor number selection via information criteria
  \item Long-run variance estimation using Bartlett kernel
}
}
\references{
Westerlund, J. (2008). Panel cointegration tests of the Fisher effect.
\emph{Journal of Applied Econometrics}, 23(2), 193--233.
\doi{10.1002/jae.967}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/muhammedalkhalaf/xtdhcoint}
  \item Report bugs at \url{https://github.com/muhammedalkhalaf/xtdhcoint/issues}
}

}
\author{
\strong{Maintainer}: Muhammad Alkhalaf \email{muhammedalkhalaf@gmail.com} (\href{https://orcid.org/0009-0002-2677-9246}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item
  \item Joakim Westerlund (Original GAUSS code author) [contributor]
}

}
\keyword{internal}
